favorite list, edit info customer, order detail
add/remove favorite list comic
edit information customer (password can't change ???)
order detail
validate email, pws part 2
fix id to Long type
add some entity created at + updated at

+++ tmanga/src/main/java/com/website/springmvc/Services/CartDetailService.java
@@ -26,7 +26,7 @@ public class CartDetailService {
 		return session.createQuery("from CartDetail where cart.id = :idCart").setParameter("idCart", idCart).list();
 	}
 	
-	public CartDetail getDetailByCartAndProduct(Long idCart, int idComic) {
+	public CartDetail getDetailByCartAndProduct(Long idCart, Long idComic) {
 		Session session = this.sessionFactory.getCurrentSession();
 		return (CartDetail) session.createQuery("from CartDetail where cart.id = :idCart and comic.id = :idComic").setParameter("idCart", idCart).setParameter("idComic", idComic).uniqueResult();
 	}
